TCREATE and TOPEN: mask open flags so buggy clients do not break Some clients will incorrectly set bits in 9p flags that don't make sense. For example, the linux 9p kernel client propagates O_DIRECT to TCREATE and TOPEN and from there to the server.
Processes on client machines set O_DIRECT to bypass the cache, but if
the server uses O_DIRECT in the open or create, then subsequent server
write and read system calls will fail, as O_DIRECT requires at minimum 512
byte aligned data, and the data is almost always not aligned.

While the linux kernel client is arguably broken, we won't be able
to fix every kernel out there, and this is surely not the only buggy client
we will see.

The fix is to enumerate the set of flags we support and then & that with
the flags received in a TCREATE or TOPEN. This nicely fixes a real problem
we are seeing with a file system benchmark.

This is in line with the general principle of being liberal with what you
accept.

Signed-off-by: Ronald G. Minnich <rminnich@gmail.com>