fix Qid size calculation
The size of Qid was being calculated using std::mem::size_of_val. This
function counts padding due to the Rust compiler aligning fields which
results in an incorrect wire size. This patch adds a size function
similar to other structures to explicitly calculate the size of Qid
without padding.